AG-9206 - Extract useJsObjectSelection as a util hook
